engine: extract esm/layers/layer.js
